no requirements.txt yet

+++ install.sh
@@ -321,11 +321,13 @@ install_zsh() {
 }
 
 
-if [ ! -e $root/.venv ]; then
-    python3 -m venv $root/.venv
-    $root/.venv pip install --upgrade setuptools pip
+if [ -e $root/requirements.txt ]; then
+    if [ ! -e $root/.venv ]; then
+        python3 -m venv $root/.venv
+        $root/.venv pip install --upgrade setuptools pip
+    fi
+    $root/.venv pip install --upgrade --requirements $root/requirements.txt
 fi
-$root/.venv pip install --upgrade --requirements $root/requirements.txt
 
 
 printf "Linking shared configs... "
